environment refactor
