+++ packages/pyright-lsif/TODO.md
@@ -1,0 +1,5 @@
+
+- [ ] Environment improvements:
+  - [ ] Allow passing environment information via json
+  - [ ] Cache environment information and read it
+  - [ ] Allow skipping environment information
