+++ packages/pyright-lsif/snapshots/input/dunder_vars/__main__.py
