+++ packages/pyright-lsif/snapshots/input/relative_import/bar.py
