+++ packages/pyright-lsif/snapshots/input/relative_import/foo.py
