--- packages/pyright-lsif/snapshots/input/simple_class/simplest.py
@@ -1,6 +1,0 @@
-class MyClass:
-    """
-    This is an example class of things that I'm making
-    """
-
-    a: int
