+++ packages/pyright-lsif/snapshots/input/single_function/src/single_function.py
@@ -1,5 +1,4 @@
 def my_cool_function(a: str) -> str:
     return a
 
-if True:
-    my_cool_function("hello")
+my_cool_function("hello")
