+++ packages/pyright-lsif/snapshots/output/dunder_vars/__main__.py
@@ -1,0 +1,7 @@
+
+if __name__ == '__main__':
+#  ^^^^^^^^ reference  python 3.10 __name__.
+    print("main")
+#   ^^^^^ reference  builtins 3.9 print().
+
+
