+++ packages/pyright-lsif/snapshots/output/f_string/fstring.py
@@ -1,7 +1,7 @@
 var = ", world!"
-#^^ definition  fstring test var.
+#^^ definition  fstring unknown var.
 
 print(f"var: hello {var}")
 #^^^^ reference  builtins 3.9 print().
-#                   ^^^ reference  fstring test var.
+#                   ^^^ reference  fstring unknown var.
