+++ packages/pyright-lsif/snapshots/output/nested_items/src/foo/bar/__init__.py
@@ -1,5 +1,5 @@
 class InitClass:
-#     ^^^^^^^^^ definition  src/foo.bar test InitClass#
+#     ^^^^^^^^^ definition  src/foo.bar unknown InitClass#
     init_item = 10
-#   ^^^^^^^^^ definition  src/foo.bar test InitClass#init_item.
+#   ^^^^^^^^^ definition  src/foo.bar unknown InitClass#init_item.
