+++ packages/pyright-lsif/snapshots/output/nested_items/src/foo/bar/baz/mod.py
@@ -1,6 +1,6 @@
 class SuchNestedMuchWow:
-#     ^^^^^^^^^^^^^^^^^ definition  src/foo.bar.baz.mod test SuchNestedMuchWow#
+#     ^^^^^^^^^^^^^^^^^ definition  src/foo.bar.baz.mod unknown SuchNestedMuchWow#
     class_item: int = 42
-#   ^^^^^^^^^^ definition  src/foo.bar.baz.mod test SuchNestedMuchWow#class_item.
+#   ^^^^^^^^^^ definition  src/foo.bar.baz.mod unknown SuchNestedMuchWow#class_item.
 #               ^^^ reference  builtins 3.9 int#
