+++ packages/pyright-lsif/snapshots/output/poetry_project/src/__init__.py
@@ -1,8 +1,8 @@
 import requests
-#      ^^^^^^^^ reference  requests test requests/__init__:
+#      ^^^^^^^^ reference  requests 2.27.1 requests/__init__:
 
 # requests.get("https://google.com")
 print(requests)
 #^^^^ reference  builtins 3.9 print().
-#     ^^^^^^^^ reference  requests test requests/__init__:
+#     ^^^^^^^^ reference  requests 2.27.1 requests/__init__:
