+++ packages/pyright-lsif/snapshots/output/single_function/src/single_function.py
@@ -1,12 +1,11 @@
 def my_cool_function(a: str) -> str:
-#   ^^^^^^^^^^^^^^^^ definition  src/single_function test my_cool_function().
-#                    ^ definition  src/single_function test my_cool_function().(a)
+#   ^^^^^^^^^^^^^^^^ definition  src/single_function unknown my_cool_function().
+#                    ^ definition  src/single_function unknown my_cool_function().(a)
 #                       ^^^ reference  builtins 3.9 str#
 #                               ^^^ reference  builtins 3.9 str#
     return a
-#          ^ reference  src/single_function test my_cool_function().(a)
+#          ^ reference  src/single_function unknown my_cool_function().(a)
 
-if True:
-    my_cool_function("hello")
-#   ^^^^^^^^^^^^^^^^ reference  src/single_function test my_cool_function().
+my_cool_function("hello")
+#^^^^^^^^^^^^^^^ reference  src/single_function unknown my_cool_function().
